backend/requirement.txt
asgiref==3.7.2
bcrypt==4.1.2
Django==5.0.3
django-cors-headers==4.3.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uuid==1.30
